requirements.txt
python-telegram-bot
requests
matplotlib
flask
python-dotenv
psutil
prometheus_client
